requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic
google-generativeai==0.3.2
google-cloud-aiplatform==1.36.0
vertexai
pytest
pytest-asyncio
requests
pytest-cov
beautifulsoup4==4.12.3